feat: add division and floor division methods to Size3D; update rotation transforms type annotations to appease mypy (#247)

+++ src/imgtools/transforms/functional.py
@@ -271,7 +271,7 @@ def zoom(
 
 def rotate(
     image: sitk.Image,
-    rotation_centre: List[float],
+    rotation_centre: List[int],
     angles: List[float],
     interpolation: str = "linear",
 ) -> sitk.Image:
